requirements.txt
openai>=1.15.0
gitpython>=3.1.43
python-dotenv>=1.0.1
